Fixed the data type of membership sign up fee and populated the membership type in the database

+++ Models/MembershipType.cs
@@ -8,7 +8,7 @@ namespace Eventy.Models
     public class MembershipType
     {
         public byte Id { get; private set; }
-        public short SignUpFee { get; set; }
+        public decimal SignUpFee { get; set; }
         public byte DurationInMonths { get; set; }
         public byte DiscountRate { get; set; }
